add only author delete question

+++ app/controllers/questions_controller.rb
@@ -32,8 +32,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'Question successfully deleted.'
+    else
+      redirect_to question_path, notice: 'Only author can delete question.'
+    end
   end
 
   private
